requirements.txt
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
appdirs==1.4.4
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
fastapi==0.110.0
filelock==3.13.3
fs==2.4.16
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6 
motor==3.4.0
openai==1.14.3
packaging==24.0
path==16.10.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
